+++ src/BlendedConcept/Disability/Presentation/HTTP/AccessibilityDeviceController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Response;
 use Inertia\Inertia;
 use Src\BlendedConcept\Disability\Application\DTO\DeviceData;
 use Src\BlendedConcept\Disability\Application\Mappers\DeviceMapper;
+use Src\BlendedConcept\Disability\Application\Policies\AccessibilityDevicePolicy;
 use Src\BlendedConcept\Disability\Application\Requests\StoreDeviceRequest;
 use Src\BlendedConcept\Disability\Application\Requests\UpdateDeviceRequest;
 use Src\BlendedConcept\Disability\Application\UseCases\Commands\Devices\DeleteDeviceCommand;
@@ -20,6 +21,8 @@ class AccessibilityDeviceController
 {
     public function index()
     {
+        abort_if(authorize('view', AccessibilityDevicePolicy::class), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $filters = request(['search', 'page', 'perPage', 'filter']);
         $devices = (new GetDevices($filters))->handle();
 
@@ -33,7 +36,8 @@ class AccessibilityDeviceController
         // dd($request->all());
         try {
             // Abort if the user is not authorized to create Devices
-            // abort_if(authorize('create', DevicePolicy::class), Response::HTTP_FORBIDDEN, '403 Forbidden');
+            abort_if(authorize('store', AccessibilityDevicePolicy::class), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
 
             // Validate the request data
 
@@ -54,6 +58,8 @@ class AccessibilityDeviceController
 
     public function create()
     {
+        abort_if(authorize('create', AccessibilityDevicePolicy::class), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $disabilityTypes = (new GetDisabilityTypeForSelect())->handle();
         $books = (new GetSimpleBooks())->handle();
 
@@ -65,6 +71,8 @@ class AccessibilityDeviceController
 
     public function edit($id)
     {
+        abort_if(authorize('edit', AccessibilityDevicePolicy::class), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $device = DeviceEloquentModel::with('books')->findOrFail($id);
         $disabilityTypes = (new GetDisabilityTypeForSelect())->handle();
         $books = (new GetSimpleBooks())->handle();
@@ -77,6 +85,7 @@ class AccessibilityDeviceController
 
     public function update(UpdateDeviceRequest $request, $id)
     {
+        abort_if(authorize('update', AccessibilityDevicePolicy::class), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         try {
 
@@ -97,6 +106,8 @@ class AccessibilityDeviceController
 
     public function destroy($id)
     {
+        abort_if(authorize('destroy', AccessibilityDevicePolicy::class), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         try {
             $device = DeviceEloquentModel::findOrFail($id);
             $deleteDeviceCommand = (new DeleteDeviceCommand($device));
