+++ src/BlendedConcept/Organisation/Presentation/HTTP/OrganisationStudentController.php
@@ -2,9 +2,11 @@
 
 namespace Src\BlendedConcept\Organisation\Presentation\HTTP;
 
+use Illuminate\Http\Response;
 use Inertia\Inertia;
 use Src\BlendedConcept\Organisation\Application\DTO\StudentData;
 use Src\BlendedConcept\Organisation\Application\Mappers\StudentMapper;
+use Src\BlendedConcept\Organisation\Application\Policies\OrganisationUserPolicy;
 use Src\BlendedConcept\Organisation\Application\Requests\StoreStudentRequest;
 use Src\BlendedConcept\Organisation\Application\Requests\UpdateStudentRequest;
 use Src\BlendedConcept\Organisation\Application\UseCases\Commands\Student\CreateStudentCommand;
@@ -19,6 +21,7 @@ class OrganisationStudentController
 {
     public function index()
     {
+        abort_if(authorize('view', OrganisationUserPolicy::class), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $filters = request(['search', 'first_name', 'last_name', 'email']) ?? [];
         $studentListWithPagniation = (new GetStudentList($filters))->handle();
 
@@ -30,6 +33,7 @@ class OrganisationStudentController
     public function create()
     {
         // $disability_types =
+        abort_if(authorize('create', OrganisationUserPolicy::class), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $learningNeeds = (new GetLearningNeed())->handle();
         $disabilityTypes = (new GetDisabilityTypes())->handle();
 
@@ -38,6 +42,7 @@ class OrganisationStudentController
 
     public function store(StoreStudentRequest $request)
     {
+        abort_if(authorize('store', OrganisationUserPolicy::class), Response::HTTP_FORBIDDEN, '403 Forbidden');
         try {
             $newStudent = StudentMapper::fromRequest($request);
             $student = (new CreateStudentCommand($newStudent))->execute();
@@ -50,7 +55,7 @@ class OrganisationStudentController
 
     public function edit(StudentEloquentModel $organisations_student)
     {
-
+        abort_if(authorize('edit', OrganisationUserPolicy::class), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $organisations_student->load(['organisation', 'user', 'disability_types', 'learningneeds', 'parent']);
         $learningNeeds = (new GetLearningNeed())->handle();
         $disability_types = (new GetDisabilityTypes())->handle();
@@ -64,7 +69,7 @@ class OrganisationStudentController
     public function update(StudentEloquentModel $organisations_student, UpdateStudentRequest $request)
     {
 
-
+        abort_if(authorize('update', OrganisationUserPolicy::class), Response::HTTP_FORBIDDEN, '403 Forbidden');
         try {
             $updateStudent = StudentData::fromRequest($request, $organisations_student);
             (new UpdateStudentCommand($updateStudent))->execute();
@@ -78,7 +83,7 @@ class OrganisationStudentController
 
     public function show(StudentEloquentModel $organisations_student)
     {
-
+        abort_if(authorize('show', OrganisationUserPolicy::class), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $organisations_student->load(['organisation', 'user', 'disability_types', 'learningneeds', 'parent']);
 
         // // return $organisations_student;
@@ -87,6 +92,7 @@ class OrganisationStudentController
 
     public function destroy(StudentEloquentModel $organisations_student)
     {
+        abort_if(authorize('destroy', OrganisationUserPolicy::class), Response::HTTP_FORBIDDEN, '403 Forbidden');
         (new DeleteStudentCommand($organisations_student))->execute();
 
         return to_route('organisations-teacher.index')->with('successMessage', 'Student Deleted Successfully!');
