+++ src/BlendedConcept/StoryBook/Presentation/HTTP/GameController.php
@@ -51,7 +51,7 @@ class GameController
      */
     public function store(StoreGameRequest $request)
     {
-        // abort_if(authorize('create', GamePolicy::class), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(authorize('create', GamePolicy::class), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         try {
             // Validate the request data
@@ -80,7 +80,7 @@ class GameController
     public function update(UpdateGameRequest $request, GameEloquentModel $game)
     {
 
-        // abort_if(authorize('edit', GamePolicy::class), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(authorize('edit', GamePolicy::class), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         /**
          * Validate the request.
@@ -107,6 +107,8 @@ class GameController
 
     public function destroy($id)
     {
+        abort_if(authorize('destroy', GamePolicy::class), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         try {
             $game = GameEloquentModel::findOrFail($id);
             $deleteGameCommand = (new DeleteGameCommand($game));
